[refactor] rename interceptors
