+++ backend/src/main/java/com/woowacourse/edd/interceptor/NoSignInRequiredInterceptor.java
@@ -1,6 +1,6 @@
 package com.woowacourse.edd.interceptor;
 
-import com.woowacourse.edd.exceptions.InvalidAccessException;
+import com.woowacourse.edd.exceptions.UnauthenticatedException;
 import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
 
 import javax.servlet.http.HttpServletRequest;
@@ -9,11 +9,11 @@ import javax.servlet.http.HttpSession;
 import java.util.Arrays;
 import java.util.List;
 
-public class MethodNeedNotSignedInInterceptor extends HandlerInterceptorAdapter {
+public class NoSignInRequiredInterceptor extends HandlerInterceptorAdapter {
 
     private final List<String> allowedMethods;
 
-    public MethodNeedNotSignedInInterceptor(String... allowedMethods) {
+    public NoSignInRequiredInterceptor(String... allowedMethods) {
         this.allowedMethods = Arrays.asList(allowedMethods);
     }
 
@@ -21,19 +21,19 @@ public class MethodNeedNotSignedInInterceptor extends HandlerInterceptorAdapter 
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
         String method = request.getMethod();
 
-        if (isNotSignedIn(request.getSession()) ||
+        if (isSignedIn(request.getSession()) ||
             shouldNotVerify(method)) {
             return true;
         }
 
-        throw new InvalidAccessException();
+        throw new UnauthenticatedException();
     }
 
     private boolean shouldNotVerify(String method) {
         return !allowedMethods.contains(method);
     }
 
-    private boolean isNotSignedIn(HttpSession session) {
-        return session.getAttribute("user") == null;
+    private boolean isSignedIn(HttpSession session) {
+        return session.getAttribute("user") != null;
     }
 }
