+++ backend/src/main/java/com/woowacourse/edd/interceptor/SignInRequiredInterceptor.java
@@ -1,6 +1,6 @@
 package com.woowacourse.edd.interceptor;
 
-import com.woowacourse.edd.exceptions.UnauthenticatedException;
+import com.woowacourse.edd.exceptions.InvalidAccessException;
 import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
 
 import javax.servlet.http.HttpServletRequest;
@@ -9,11 +9,11 @@ import javax.servlet.http.HttpSession;
 import java.util.Arrays;
 import java.util.List;
 
-public class MethodNeedSignedInInterceptor extends HandlerInterceptorAdapter {
+public class SignInRequiredInterceptor extends HandlerInterceptorAdapter {
 
     private final List<String> allowedMethods;
 
-    public MethodNeedSignedInInterceptor(String... allowedMethods) {
+    public SignInRequiredInterceptor(String... allowedMethods) {
         this.allowedMethods = Arrays.asList(allowedMethods);
     }
 
@@ -21,19 +21,19 @@ public class MethodNeedSignedInInterceptor extends HandlerInterceptorAdapter {
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
         String method = request.getMethod();
 
-        if (isSignedIn(request.getSession()) ||
+        if (isNotSignedIn(request.getSession()) ||
             shouldNotVerify(method)) {
             return true;
         }
 
-        throw new UnauthenticatedException();
+        throw new InvalidAccessException();
     }
 
     private boolean shouldNotVerify(String method) {
         return !allowedMethods.contains(method);
     }
 
-    private boolean isSignedIn(HttpSession session) {
-        return session.getAttribute("user") != null;
+    private boolean isNotSignedIn(HttpSession session) {
+        return session.getAttribute("user") == null;
     }
 }
